Simplify ResourceCard, accept onUpdated function from parent
